func-tests: GetHCO to non ginkgo func
The GetHCO helper function is now gomega free, and does not use
assersions. Instead, it return an error to be handled by the caller.
