+++ tests/func-tests/monitoring_test.go
@@ -168,7 +168,8 @@ var _ = Describe("[crit:high][vendor:cnv-qe@redhat.com][level:system]Monitoring"
 	It("UnsupportedHCOModification alert should fired when there is a jsonpatch annotation to modify an operand CRs", func(ctx context.Context) {
 		query := fmt.Sprintf(`kubevirt_hco_unsafe_modifications{annotation_name=%q}`, common.JSONPatchKVAnnotationName)
 
-		hco := tests.GetHCO(ctx, cli)
+		hco, err := tests.GetHCO(ctx, cli)
+		Expect(err).ToNot(HaveOccurred())
 
 		hco.Annotations = map[string]string{
 			common.JSONPatchKVAnnotationName: `[{"op": "add", "path": "/spec/configuration/migrations", "value": {"allowPostCopy": true}}]`,
@@ -195,10 +196,17 @@ var _ = Describe("[crit:high][vendor:cnv-qe@redhat.com][level:system]Monitoring"
 		}).WithTimeout(prometheousTimeout).
 			WithPolling(prometheousPolling).
 			WithContext(ctx).
-			ShouldNot(BeNil(), tests.PrintHyperConvergedBecause(
-				tests.GetHCO(ctx, cli),
-				`can't find the "UnsupportedHCOModification" alert in the list of firing alerts: %s`, getAlertNames(alerts),
-			))
+			ShouldNot(BeNil(), func() func() string {
+				hc, err := tests.GetHCO(ctx, cli)
+				if err != nil {
+					return func() string {
+						return "Error reading HyperConverged CR; " + err.Error()
+					}
+				}
+				return tests.PrintHyperConvergedBecause(hc,
+					`can't find the "UnsupportedHCOModification" alert in the list of firing alerts: %s`, getAlertNames(alerts),
+				)
+			})
 	})
 
 	It("should fire the DeprecatedMachineType alert when a VM is using a deprecated machine type", func(ctx context.Context) {
