+++ tests/func-tests/utils.go
@@ -6,7 +6,6 @@ import (
 	"errors"
 	"flag"
 	"fmt"
-	"slices"
 	"sync"
 	"time"
 
@@ -17,7 +16,6 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -25,11 +23,6 @@ import (
 	"k8s.io/client-go/tools/remotecommand"
 	"k8s.io/utils/ptr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	hcov1 "github.com/kubevirt/hyperconverged-cluster-operator/api/v1"
-	"github.com/kubevirt/hyperconverged-cluster-operator/api/v1/featuregates"
-	hcov1beta1 "github.com/kubevirt/hyperconverged-cluster-operator/api/v1beta1"
-	hcoutil "github.com/kubevirt/hyperconverged-cluster-operator/pkg/util"
 )
 
 var (
@@ -133,122 +126,6 @@ func IsOpenShift(ctx context.Context, cli client.Client) (bool, error) {
 	return isOpenShiftCache.IsOpenShift(ctx, cli)
 }
 
-// GetHCO reads the HCO CR from the APIServer with a DynamicClient
-func GetHCO(ctx context.Context, cli client.Client) *hcov1.HyperConverged {
-	ginkgo.GinkgoHelper()
-
-	hco := &hcov1.HyperConverged{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      hcoutil.HyperConvergedName,
-			Namespace: InstallNamespace,
-		},
-	}
-
-	Expect(cli.Get(ctx, client.ObjectKeyFromObject(hco), hco)).To(Succeed())
-
-	return hco
-}
-
-// GetHCOv1beta1 reads the HCO CR from the APIServer with a DynamicClient
-func GetHCOv1beta1(ctx context.Context, cli client.Client) *hcov1beta1.HyperConverged {
-	ginkgo.GinkgoHelper()
-
-	hco := &hcov1beta1.HyperConverged{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      hcoutil.HyperConvergedName,
-			Namespace: InstallNamespace,
-		},
-	}
-
-	Expect(cli.Get(ctx, client.ObjectKeyFromObject(hco), hco)).To(Succeed())
-
-	return hco
-}
-
-// UpdateHCORetry updates the HCO CR in a safe way internally calling UpdateHCO_old
-// UpdateHCORetry internally uses an async Eventually block refreshing the in-memory
-// object if needed and setting there Spec, Annotations, Finalizers and Labels from the
-// input object.
-// UpdateHCORetry should be preferred over UpdateHCO_old to reduce test flakiness due to
-// inevitable concurrency conflicts
-func UpdateHCORetry(ctx context.Context, cli client.Client, input *hcov1.HyperConverged) *hcov1.HyperConverged {
-	var output *hcov1.HyperConverged
-	var err error
-
-	Eventually(func(ctx context.Context) error {
-		hco := GetHCO(ctx, cli)
-		input.Spec.DeepCopyInto(&hco.Spec)
-		hco.ObjectMeta.Annotations = input.ObjectMeta.Annotations
-		hco.ObjectMeta.Finalizers = input.ObjectMeta.Finalizers
-		hco.ObjectMeta.Labels = input.ObjectMeta.Labels
-
-		output, err = UpdateHCO(ctx, cli, hco)
-		return err
-	}).WithTimeout(time.Minute).WithPolling(10 * time.Second).WithContext(ctx).Should(Succeed())
-
-	return output
-}
-
-// UpdateHCO updates the HCO CR using a DynamicClient, it can return errors on failures
-func UpdateHCO(ctx context.Context, cli client.Client, input *hcov1.HyperConverged) (*hcov1.HyperConverged, error) {
-	err := hcov1.AddToScheme(cli.Scheme())
-	if err != nil {
-		return nil, err
-	}
-
-	hco := GetHCO(ctx, cli)
-	input.Spec.DeepCopyInto(&hco.Spec)
-	hco.Annotations = input.Annotations
-	hco.Finalizers = input.Finalizers
-	hco.Labels = input.Labels
-	hco.Status = hcov1.HyperConvergedStatus{} // to silence warning about unknown fields.
-
-	err = cli.Update(ctx, hco)
-	if err != nil {
-		return nil, err
-	}
-
-	hco = GetHCO(ctx, cli)
-	return hco, nil
-}
-
-// PatchHCO updates the HCO CR using a DynamicClient, it can return errors on failures
-func PatchHCO(ctx context.Context, cli client.Client, patchBytes []byte) error {
-	patch := client.RawPatch(types.JSONPatchType, patchBytes)
-	hco := &hcov1.HyperConverged{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      hcoutil.HyperConvergedName,
-			Namespace: InstallNamespace,
-		},
-	}
-
-	return cli.Patch(ctx, hco, patch)
-}
-
-// PatchMergeHCO patches the HCO CR using a DynamicClient, it can return errors on failures
-func PatchMergeHCO(ctx context.Context, cli client.Client, patchBytes []byte) error {
-	patch := client.RawPatch(types.MergePatchType, patchBytes)
-	hco := &hcov1.HyperConverged{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      hcoutil.HyperConvergedName,
-			Namespace: InstallNamespace,
-		},
-	}
-
-	return cli.Patch(ctx, hco, patch)
-}
-
-func RestoreDefaults(ctx context.Context, cli client.Client) {
-	Eventually(func(ctx context.Context) error {
-		return PatchHCO(ctx, cli, []byte(`[{"op": "replace", "path": "/spec", "value": {}}]`))
-	}).
-		WithOffset(1).
-		WithTimeout(time.Second * 5).
-		WithPolling(time.Millisecond * 100).
-		WithContext(ctx).
-		Should(Succeed())
-}
-
 func deleteAllResources(ctx context.Context, restClient rest.Interface, resourceName string) {
 	Eventually(func() bool {
 		err := restClient.Delete().Namespace(TestNamespace).Resource(resourceName).Do(ctx).Error()
@@ -332,28 +209,4 @@ func executeCommandOnPod(ctx context.Context, k8scli *kubernetes.Clientset, pod 
 		return "", "", fmt.Errorf("%w Failed executing command %s on %v/%v", err, command, pod.Namespace, pod.Name)
 	}
 	return buf.String(), errBuf.String(), nil
-}
-
-func EnableFG(ctx context.Context, cli client.Client, fgName string) error {
-	hc := GetHCO(ctx, cli)
-	if hc.Spec.FeatureGates == nil {
-		patch := fmt.Appendf(nil, `{"spec":{"featureGates":[{"name": %q, "state": "%v"}]}}`, fgName, featuregates.Enabled)
-		return PatchMergeHCO(ctx, cli, patch)
-	}
-
-	if !hc.Spec.FeatureGates.IsEnabled(fgName) {
-		var patch []byte
-		idx := slices.IndexFunc(hc.Spec.FeatureGates, func(fg featuregates.FeatureGate) bool {
-			return fg.Name == fgName
-		})
-		if idx == -1 {
-			patch = fmt.Appendf(nil, `[{"op": "add", "path": "/spec/featureGates/-", "value": {"name": %q}}]`, fgName)
-		} else {
-			patch = fmt.Appendf(nil, `[{"op": "replace", "path": "/spec/featureGates/%d/state", "value": "%v"}]`, idx, featuregates.Enabled)
-		}
-
-		return PatchHCO(ctx, cli, patch)
-	}
-
-	return nil
 }
